Modified Struct Definitions

+++ Part1/block.hpp
@@ -4,13 +4,13 @@
 #include "transaction.h"
 
 struct Block {
-    int id;
-    int height;
-    int parent_id;
+    uint64_t id;
+    uint64_t height;
+    uint64_t parent_id;
     std::vector<Transaction> transactions;
     time_t timestamp;
 
-    Block(int id, int height, int parent_id, std::vector<Transaction> transactions, time_t timestamp):
+    Block(uint64_t id, uint64_t height, uint64_t parent_id, std::vector<Transaction> transactions, time_t timestamp):
         id(id),
         height(height),
         parent_id(parent_id),
@@ -18,7 +18,7 @@ struct Block {
         timestamp(timestamp)
     {}
 
-    Block(int id, int height, int parent_id, time_t timestamp):
+    Block(uint64_t id, uint64_t height, uint64_t parent_id, time_t timestamp):
         id(id),
         height(height),
         parent_id(parent_id),
